src/requirements.txt
# Core Dependencies
langchain>=0.2.0
langchain-core>=0.2.0
langgraph>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# MCP Protocol
mcp>=0.1.0

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0

# Audio/Video Processing
ffmpeg-python>=0.2.0
openai-whisper>=20231117
moviepy>=1.0.3

# Image/Video Enhancement
opencv-python>=4.8.0
Pillow>=10.0.0

# Object Detection (YOLO)
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Utilities
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiofiles>=23.0.0

# UI (Optional)
PyQt6>=6.6.0

# Vector Database
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment if using GPU

# Monitoring
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
